feat: keep dnd handler in separate component
